Day 2 Commit
I decided that I would be needing more variables, and hence included them. I will need to set a timer, either from high to low or low to high, but I’ll decide that later. I also added a background image, but I’m not sure how to check if it looks right, so I will be doing that later as well, when I have at least part of my main code.