Day 3 Commit
I coded for a title screen that will stay till the user clicks anywhere within the pygame screen, which contains simple rules and instructions of the game. I added some boolean variables that will help me with the main loop of my code, and coded Class for three of the in-game “characters” that I have so far. I believe I will change the images later on because I don’t think it pairs well with the background, but that will be for later (might draw it myself).