Day 4 Commit
I took a good break and thought about what I should do to create the eating motion of the game. I’m not sure how to find it online (I’m probably missing out on the key words), but it seems I have to draw it myself. I’ll see how that goes. 
Goal for tonight: Finish drawing the motions for both in game characters: bunny and fox. 
Goal for next class: I want to be able to attach an image of the characters in the title screen, as well as have the characters seated in a designated position on the background image.