Add files via upload
I added the pancakes and waffles that will be the only in game variables that will be able to move, ideally. I coded for a conditional statement that will detect which (pancake or waffle) item the character has "eaten" to deduct or add the correct points towards variables pancake and or waffle.