Day 5 Commit
Regarding what I said about the drawings, once again, I did not finish the character drawings, but don’t worry, I will get it done sooner or later of course. I am coding for the movement of the pancake, waffle, and paffle today, and ideally figure out how to get the characters (to be replaced) to sit in one spot without moving but eat the food items when colliding with them.
I finished coding for the movement of the pancake, waffle, and paffle, but I did not achieve the second part. Instead, I coded for the conditions that would cause a certain message to appear. 
Goal for next class: Get the main screen running.