Day 6 Commit
I readjusted the sizing of the images till it was to my liking, and I picked out a new background because the previous background did not fit right. The main screen is running now, so my new goal is to code for detection of collision between the food items and the characters to update specific variables such as waffles_eaten and pancakes_eaten.