Day 7 Commit
I coded for the detection of collision between the food items and the characters, but now I have to figure out how to update the number for the display of pancakes_eaten and waffles_eaten. I managed to get the number to update, but it is updating at an interestingly high speed, hence I will have to fix that.