Day 9 Commit
I adjusted the features of the game more to my liking, and I fixed the point value of variables pancakes_eaten and waffles_eaten so that it is no longer going up by an abnormal amount. My next goal is to draw more images so that there will be an eating motion for the characters when collided with the food item.