Day 11 Commit
I have decided that when shown != paffle, paffle will be floating somewhere in the air just for visual purposes. That will be the speciality of a paffle, hence, I have finished coding for the paffle. I have also decided that it would be really sad to lose value in the food items, and it seems rather weird if I tie it to real world situations, hence there will be no losing food in this game. 
My goal for tomorrow and Friday:
5/30: find a carton of milk pixelated image and search on the internet to find out how to get an item to fall from the top of the screen
5/31: finish coding for the image added on 5/30 and overall code