Day 14 Commit
I found an image of ditto and decided to code for that as the object to be dragged. I think ditto is a very cool choice because ditto can morph into anything, and that’s a great idea for coding, but I am not skilled enough for that as of right now. Though I have not successfully written the code for it to be dragged yet, I will be able to with some fixes soon.
My goal for next class or tomorrow (6/4) will be to code for ditto to be dragged within the game screen.