Day 15 Commit
I tried to code for the ditto to be dragged numerous times but the result was either that my game would not run, the food items would not move, or the food items moved really slow. I will work on this over the weekend to have more time to think.